requirements.txt
playwright==1.44.0
gspread==6.1.2
google-auth==2.29.0
requests==2.31.0
